[Refactor] #189 - 뷰 뷰모델 분리
